+++ Lib/TSCompiler/FastBundleBundler.cs
@@ -118,9 +118,9 @@ namespace Lib.TSCompiler
             if (testProj)
             {
                 FilesContent["test.html"] = _indexHtml;
-                FilesContent["jasmine-core.js"] = _tools.JasmineCoreJs;
-                FilesContent["jasmine-boot.js"] = _tools.JasmineBootJs;
-                FilesContent["loader.js"] = _tools.LoaderJs;
+                FilesContent[_versionDirPrefix + "jasmine-core.js"] = _tools.JasmineCoreJs;
+                FilesContent[_versionDirPrefix + "jasmine-boot.js"] = _tools.JasmineBootJs;
+                FilesContent[_versionDirPrefix + "loader.js"] = _tools.LoaderJs;
             }
             else
             {
@@ -146,14 +146,14 @@ namespace Lib.TSCompiler
         <title>{Project.Title}</title>{cssLink}
     </head>
     <body>{InitG11n()}
-        <script src=""jasmine-core.js"" charset=""utf-8""></script>
-        <script src=""jasmine-boot.js"" charset=""utf-8""></script>
-        <script src=""loader.js"" charset=""utf-8""></script>
+        <script src=""{_versionDirPrefix}jasmine-core.js"" charset=""utf-8""></script>
+        <script src=""{_versionDirPrefix}jasmine-boot.js"" charset=""utf-8""></script>
+        <script src=""{_versionDirPrefix}loader.js"" charset=""utf-8""></script>
         <script>
             {GetGlobalDefines()}
             {GetModuleMap()}
         </script>
-        <script src=""testbundle.js"" charset=""utf-8""></script>
+        <script src=""{_versionDirPrefix}testbundle.js"" charset=""utf-8""></script>
         <script>
             {RequireBobril()} {reqSpec}
         </script>